site/requirements.txt
paho-mqtt==1.5.1
Flask==2.0.1
gunicorn==20.0.4